MNT: Move Document to commands.py.
This is a necessary step so that event_descriptors can be
properly dereferenced for Events.  I need to use _find_latest()
in _AsDocument.from_dict() but mds.document cannot import
mds.commands since mds.commands was importing mds.document.
